fixed service reference errors with reorganization of frontend
