+++ src/app/components/admin-inventory-page/admin-inventory-page.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { InventoryItem, InventoryItemsService } from '../../services/inventory-items.service'
-import { HttpUserInventoryPageService } from '../user-inventory-page/http-user-inventory-page.service';
+import { HttpUserInventoryPageService } from '../../services/http-user-inventory-page.service';
 
 @Component({
   selector: 'app-admin-inventory-page',
